backend/requirements.txt
# ============================================================
# BookAutomation Backend Dependencies
# ============================================================

# Django & REST Framework
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9

# Celery & Redis (Upstash compatible)
celery==5.4.0
redis==5.1.0
django-celery-results==2.5.1
django-celery-beat==2.7.0
django-redis==5.4.0

# Vector Database (Using pgvector on Neon)
pgvector==0.3.6
numpy==2.2.3

# AI - Google Gemini
google-generativeai==0.8.3

# Web Scraping - Selenium + BeautifulSoup
selenium==4.25.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
requests==2.32.3

# Utilities
python-dotenv==1.0.1
Pillow==10.4.0
httpx==0.27.2

# Production
gunicorn==22.0.0
whitenoise==6.7.0
